pyproject.toml
[project]
name = "db-sample-apps"
version = "0.1.0"
description = "CLI Apps with MySQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "mysql-connector-python>=9.5.0",
    "numpy>=2.3.5",
]
